Move ssh profiles to roaming, rename sshshellprofile to sshprofile
