+++ FluentTerminal.App.Services/ApplicationDataContainers.cs
@@ -7,6 +7,6 @@
         public IApplicationDataContainer Themes { get; set; }
         public IApplicationDataContainer KeyBindings { get; set; }
         public IApplicationDataContainer ShellProfiles { get; set; }
-        public IApplicationDataContainer SshShellProfiles { get; set; }
+        public IApplicationDataContainer SshProfiles { get; set; }
     }
 }
