+++ FluentTerminal.App.Services/Constants.cs
@@ -5,7 +5,7 @@
         public const string ThemesContainerName = "Themes";
         public const string KeyBindingsContainerName = "KeyBindings";
         public const string ShellProfilesContainerName = "ShellProfiles";
-        public const string SshShellProfilesContainerName = "SshShellProfiles";
+        public const string SshProfilesContainerName = "SshProfiles";
         public const byte TerminalBufferRequestIdentifier = 0;
     }
 }
