+++ FluentTerminal.App.Services/Dialogs/ISshProfileSelectionDialog.cs
@@ -1,0 +1,12 @@
+﻿using FluentTerminal.Models;
+using System.Threading.Tasks;
+
+namespace FluentTerminal.App.Services.Dialogs
+{
+    public interface ISshProfileSelectionDialog
+    {
+        SshProfile SelectedProfile { get; }
+
+        Task<SshProfile> SelectProfile();
+    }
+}
