--- FluentTerminal.App.Services/Dialogs/ISshShellProfileSelectionDialog.cs
@@ -1,12 +1,0 @@
-﻿using FluentTerminal.Models;
-using System.Threading.Tasks;
-
-namespace FluentTerminal.App.Services.Dialogs
-{
-    public interface ISshShellProfileSelectionDialog
-    {
-        SshShellProfile SelectedProfile { get; }
-
-        Task<SshShellProfile> SelectProfile();
-    }
-}
