+++ FluentTerminal.App.Services/ISettingsService.cs
@@ -11,25 +11,25 @@ namespace FluentTerminal.App.Services
         event EventHandler KeyBindingsChanged;
         event EventHandler<ShellProfile> ShellProfileAdded;
         event EventHandler<Guid> ShellProfileDeleted;
-        event EventHandler<SshShellProfile> SshShellProfileAdded;
-        event EventHandler<Guid> SshShellProfileDeleted;
+        event EventHandler<SshProfile> SshProfileAdded;
+        event EventHandler<Guid> SshProfileDeleted;
         event EventHandler<TerminalOptions> TerminalOptionsChanged;
         event EventHandler<Guid> ThemeDeleted;
         event EventHandler<TerminalTheme> ThemeAdded;
 
         void DeleteShellProfile(Guid id);
-        void DeleteSshShellProfile(Guid id);
+        void DeleteSshProfile(Guid id);
         void DeleteTheme(Guid id);
         ApplicationSettings GetApplicationSettings();
         TerminalTheme GetCurrentTheme();
         Guid GetCurrentThemeId();
         ShellProfile GetDefaultShellProfile();
-        SshShellProfile GetDefaultSshShellProfile();
+        SshProfile GetDefaultSshProfile();
         Guid GetDefaultShellProfileId();
-        Guid GetDefaultSshShellProfileId();
+        Guid GetDefaultSshProfileId();
         IDictionary<string, ICollection<KeyBinding>> GetCommandKeyBindings();
         IEnumerable<ShellProfile> GetShellProfiles();
-        IEnumerable<SshShellProfile> GetSshShellProfiles();
+        IEnumerable<SshProfile> GetSshProfiles();
         IEnumerable<TabTheme> GetTabThemes();
         TerminalOptions GetTerminalOptions();
         TerminalTheme GetTheme(Guid id);
@@ -38,14 +38,14 @@ namespace FluentTerminal.App.Services
         void SaveApplicationSettings(ApplicationSettings applicationSettings);
         void SaveCurrentThemeId(Guid id);
         void SaveDefaultShellProfileId(Guid id);
-        void SaveDefaultSshShellProfileId(Guid id);
+        void SaveDefaultSshProfileId(Guid id);
         void SaveKeyBindings(string command, ICollection<KeyBinding> keyBindings);
         void SaveShellProfile(ShellProfile shellProfile, bool newShell = false);
-        void SaveSshShellProfile(SshShellProfile shellProfile, bool newShell = false);
+        void SaveSshProfile(SshProfile shellProfile, bool newShell = false);
         void SaveTerminalOptions(TerminalOptions terminalOptions);
         void SaveTheme(TerminalTheme theme, bool newTheme = false);
         ShellProfile GetShellProfile(Guid id);
-        SshShellProfile GetSshShellProfile(Guid id);
+        SshProfile GetSshProfile(Guid id);
         string ExportSettings();
         void ImportSettings(string serializedSettings);
     }
