+++ FluentTerminal.App.Services/Implementation/DialogService.cs
@@ -12,19 +12,19 @@ namespace FluentTerminal.App.Services.Implementation
         private readonly Func<ICreateKeyBindingDialog> _createKeyBindingDialogFactory;
         private readonly Func<IInputDialog> _inputDialogFactory;
         private readonly Func<ISshConnectionInfoDialog> _sshConnectionInfoDialogFactory;
-        private readonly Func<ISshShellProfileSelectionDialog> _sshShellProfileSelectionDialogFactory;
+        private readonly Func<ISshProfileSelectionDialog> _sshProfileSelectionDialogFactory;
 
         public DialogService(Func<IShellProfileSelectionDialog> shellProfileSelectionDialogFactory,
             Func<IMessageDialog> messageDialogFactory, Func<ICreateKeyBindingDialog> createKeyBindingDialogFactory,
             Func<IInputDialog> inputDialogFactory, Func<ISshConnectionInfoDialog> sshConnectionInfoDialogFactory,
-            Func<ISshShellProfileSelectionDialog> sshShellProfileSelectionDialogFactory)
+            Func<ISshProfileSelectionDialog> sshProfileSelectionDialogFactory)
         {
             _shellProfileSelectionDialogFactory = shellProfileSelectionDialogFactory;
             _messageDialogFactory = messageDialogFactory;
             _createKeyBindingDialogFactory = createKeyBindingDialogFactory;
             _inputDialogFactory = inputDialogFactory;
             _sshConnectionInfoDialogFactory = sshConnectionInfoDialogFactory;
-            _sshShellProfileSelectionDialogFactory = sshShellProfileSelectionDialogFactory;
+            _sshProfileSelectionDialogFactory = sshProfileSelectionDialogFactory;
         }
 
         public Task<KeyBinding> ShowCreateKeyBindingDialog()
@@ -81,9 +81,9 @@ namespace FluentTerminal.App.Services.Implementation
         public Task<ISshConnectionInfo> ShowSshConnectionInfoDialogAsync(ISshConnectionInfo input) =>
             _sshConnectionInfoDialogFactory().GetSshConnectionInfoAsync(input);
 
-        public Task<SshShellProfile> ShowSshProfileSelectionDialogAsync()
+        public Task<SshProfile> ShowSshProfileSelectionDialogAsync()
         {
-            var dialog = _sshShellProfileSelectionDialogFactory();
+            var dialog = _sshProfileSelectionDialogFactory();
 
             return dialog.SelectProfile();
         }
