+++ FluentTerminal.App.Services/Terminal.cs
