+++ FluentTerminal.App/Dialogs/SshProfileSelectionDialog.xaml
@@ -1,0 +1,19 @@
+﻿<ContentDialog
+    x:Class="FluentTerminal.App.Dialogs.SshProfileSelectionDialog"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    x:Uid="SelectSshProfileUI"
+    Title="Select SSH Profile"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
+    DefaultButton="Primary"
+    PrimaryButtonText="Select"
+    SecondaryButtonText="Cancel"
+    mc:Ignorable="d">
+    <ComboBox
+        HorizontalAlignment="Stretch"
+        DisplayMemberPath="Name"
+        ItemsSource="{x:Bind Profiles}"
+        SelectedItem="{x:Bind SelectedProfile, Mode=TwoWay}" />
+</ContentDialog>
