--- FluentTerminal.App/Dialogs/SshShellProfileSelectionDialog.xaml
@@ -1,19 +1,0 @@
-﻿<ContentDialog
-    x:Class="FluentTerminal.App.Dialogs.SshShellProfileSelectionDialog"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    x:Uid="SelectSshProfileUI"
-    Title="Select SSH Profile"
-    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
-    PrimaryButtonText="Select"
-    SecondaryButtonText="Cancel"
-    DefaultButton="Primary"
-    mc:Ignorable="d">
-    <ComboBox
-        HorizontalAlignment="Stretch"
-        DisplayMemberPath="Name"
-        ItemsSource="{x:Bind Profiles}"
-        SelectedItem="{x:Bind SelectedProfile, Mode=TwoWay}" />
-</ContentDialog>
