+++ FluentTerminal.App/Views/SettingsPages/SshProfileSettings.xaml.cs
@@ -4,7 +4,6 @@ using Windows.UI.Xaml.Navigation;
 
 namespace FluentTerminal.App.Views.SettingsPages
 {
-
     public sealed partial class SshProfileSettings : Page
     {
         public SshProfilesPageViewModel ViewModel { get; private set; }
@@ -13,14 +12,15 @@ namespace FluentTerminal.App.Views.SettingsPages
         {
             this.InitializeComponent();
         }
+
         protected override void OnNavigatedTo(NavigationEventArgs e)
         {
             if (e.Parameter is SshProfilesPageViewModel viewModel)
             {
                 ViewModel = viewModel;
-                if (ViewModel.SshShellProfiles.Count == 0)
+                if (ViewModel.SshProfiles.Count == 0)
                 {
-                    ViewModel.CreateSshShellProfile();
+                    ViewModel.CreateSshProfile();
                 }
             }
         }
